migration camera components to svelte
